create modles/canlde.go
